requirements.txt
# Django Core
Django==5.0.1
python-decouple==3.8

# Database
mysqlclient==2.2.1

# Image handling
Pillow==10.2.0

# Excel/CSV handling
openpyxl==3.1.2
pandas==2.2.0

# API (if needed)
djangorestframework==3.14.0

# Real-time features (for chat)
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Timezone handling
pytz==2024.1

# PDF generation (for reports)
reportlab==4.0.9
weasyprint==60.2

# QR Code (for attendance)
qrcode==7.4.2
Pillow==10.2.0

# Task scheduling (for notifications)
celery==5.3.6
redis==5.0.1

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.4
pytest-django==4.7.0

# Security
django-cors-headers==4.3.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2

requests==2.32.3



#pip install -r requirements.txt